Added remote debug shell
Added a simple remote debugging shell which supplants the debug app we
were using previously. This one allows us to connect observer as well.

I had to update project node to handle this, as it assumed that _any_
nodedown message was coming from the project node.
