+++ .iex.namespaced.exs
@@ -1,0 +1,1 @@
+use LXical.Server.IEx.Helpers
