--- rel/debug/env.sh.eex
@@ -1,5 +1,0 @@
-#!/bin/sh
-
-# for connecting locally
-export RELEASE_COOKIE="lexical"
-export RELEASE_DISTRIBUTION=name
