--- rel/debug/remote.vm.args.eex
@@ -1,8 +1,0 @@
-## Customize flags given to the VM: https://www.erlang.org/doc/man/erl.html
-## -mode/-name/-sname/-setcookie are configured via env vars, do not set them here
-
-## Increase number of concurrent ports/sockets
-##+Q 65536
-
-## Tweak GC to run more often
-##-env ERL_FULLSWEEP_AFTER 10
